@@ -1,14 +1,15 @@
 package axal25.oles.jacek.jdbc.dao;
 
+import axal25.oles.jacek.dao.EntityManagerTestUtils;
 import axal25.oles.jacek.entity.ApplicationEntity;
 import axal25.oles.jacek.entity.factory.ApplicationEntityFactory;
-import axal25.oles.jacek.entity.factory.EntityFactory;
-import axal25.oles.jacek.jdbc.DatabaseUtils;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.parallel.Execution;
+import org.junit.jupiter.api.parallel.ExecutionMode;
 import org.springframework.boot.test.context.SpringBootTest;
 
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.util.List;
@@ -17,64 +18,70 @@ import java.util.Random;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
+import static axal25.oles.jacek.dao.EntityManagerTestUtils.OnSuccess.ROLLBACK;
+import static axal25.oles.jacek.entity.factory.EntityFactory.IdGenerateMode.FROM_ENTITY_MANAGER;
 import static com.google.common.truth.Truth.assertThat;
 
+@Execution(ExecutionMode.SAME_THREAD)
 @SpringBootTest
 public class JdbcApplicationDaoTest {
     private static final Random random = new Random();
-    private Connection connection;
-
-    @BeforeEach
-    void setUp() throws SQLException {
-        connection = DatabaseUtils.getConnection();
-    }
-
-    @AfterEach
-    void tearDown() throws SQLException {
-        connection.rollback();
-    }
+    @PersistenceContext
+    private EntityManager entityManager;
 
     @Test
-    public void selectApplications() throws Exception {
+    public void selectApplications() {
         ApplicationEntity inputApp = ApplicationEntityFactory.produce(
                 "selectApplications",
                 getClass(),
                 null,
-                EntityFactory.IdGenerateMode.FROM_JDBC);
+                FROM_ENTITY_MANAGER);
 
-        insertApplicationAndAssert(inputApp);
-
-        List<ApplicationEntity> selecteds = JdbcApplicationDao.selectApplications(connection);
+        List<ApplicationEntity> selecteds = EntityManagerTestUtils.useHibernateConnection(
+                entityManager,
+                connection -> {
+                    insertApplicationAndAssert(inputApp, connection);
+                    return JdbcApplicationDao.selectApplications(connection);
+                },
+                ROLLBACK);
 
         assertThat(selecteds).containsAtLeastElementsIn(List.of(inputApp));
     }
 
     @Test
-    public void selectApplicationIds() throws Exception {
+    public void selectApplicationIds() {
         ApplicationEntity inputApp = ApplicationEntityFactory.produce(
                 "selectApplicationIds",
                 getClass(),
                 null,
-                EntityFactory.IdGenerateMode.FROM_JDBC);
+                FROM_ENTITY_MANAGER);
 
-        insertApplicationAndAssert(inputApp);
-
-        List<Integer> selecteds = JdbcApplicationDao.selectApplicationIds(connection);
+        List<Integer> selecteds = EntityManagerTestUtils.useHibernateConnection(
+                entityManager,
+                connection -> {
+                    insertApplicationAndAssert(inputApp, connection);
+                    return JdbcApplicationDao.selectApplicationIds(connection);
+                },
+                ROLLBACK);
 
         assertThat(selecteds).containsAtLeastElementsIn(List.of(inputApp.getId()));
     }
 
     @Test
-    public void selectApplicationById() throws Exception {
+    public void selectApplicationById() {
         ApplicationEntity inputApp = ApplicationEntityFactory.produce(
                 "selectApplicationById",
                 getClass(),
                 null,
-                EntityFactory.IdGenerateMode.FROM_JDBC);
+                FROM_ENTITY_MANAGER);
 
-        insertApplicationAndAssert(inputApp);
-
-        Optional<ApplicationEntity> selected = JdbcApplicationDao.selectApplicationById(inputApp.getId(), connection);
+        Optional<ApplicationEntity> selected = EntityManagerTestUtils.useHibernateConnection(
+                entityManager,
+                connection -> {
+                    insertApplicationAndAssert(inputApp, connection);
+                    return JdbcApplicationDao.selectApplicationById(inputApp.getId(), connection);
+                },
+                ROLLBACK);
 
         assertThat(selected).isNotNull();
         assertThat(selected.isPresent()).isTrue();
@@ -82,46 +89,56 @@ public class JdbcApplicationDaoTest {
     }
 
     @Test
-    public void selectApplicationsByIds() throws Exception {
+    public void selectApplicationsByIds() {
         List<ApplicationEntity> inputApps = IntStream.range(0, random.nextInt(10))
                 .mapToObj(unused -> ApplicationEntityFactory.produce(
                         "selectApplicationsByIds",
                         getClass(),
                         null,
-                        EntityFactory.IdGenerateMode.FROM_JDBC))
+                        FROM_ENTITY_MANAGER))
                 .collect(Collectors.toList());
 
-        inputApps.forEach(inputApp -> {
-            try {
-                insertApplicationAndAssert(inputApp);
-            } catch (SQLException e) {
-                throw new RuntimeException(e);
-            }
-        });
-
-        List<ApplicationEntity> selecteds = JdbcApplicationDao.selectApplicationsByIds(
-                inputApps.stream().map(ApplicationEntity::getId).collect(Collectors.toList()),
-                connection);
+        List<ApplicationEntity> selecteds = EntityManagerTestUtils.useHibernateConnection(
+                entityManager,
+                connection -> {
+                    inputApps.forEach(inputApp -> {
+                        try {
+                            insertApplicationAndAssert(inputApp, connection);
+                        } catch (SQLException e) {
+                            throw new RuntimeException(e);
+                        }
+                    });
+                    return JdbcApplicationDao.selectApplicationsByIds(
+                            inputApps.stream().map(ApplicationEntity::getId).collect(Collectors.toList()),
+                            connection);
+                },
+                ROLLBACK);
 
         assertThat(selecteds).containsExactlyElementsIn(inputApps);
     }
 
     @Test
-    public void insertApplication_isInsertedSuccessfully() throws Exception {
+    public void insertApplication_isInsertedSuccessfully() {
         ApplicationEntity inputApp = ApplicationEntityFactory.produce(
                 "insertApplication_isInsertedSuccessfully",
                 getClass(),
                 null,
-                EntityFactory.IdGenerateMode.FROM_JDBC);
+                FROM_ENTITY_MANAGER);
 
-        insertApplicationAndAssert(inputApp);
+        EntityManagerTestUtils.useHibernateConnection(
+                entityManager,
+                connection -> insertApplicationAndAssert(inputApp, connection),
+                ROLLBACK);
     }
 
-    private void insertApplicationAndAssert(ApplicationEntity inputApp) throws SQLException {
-        Optional<ApplicationEntity> inserted = JdbcApplicationDao.insertApplication(inputApp, connection);
+    private Void insertApplicationAndAssert(ApplicationEntity inputApp, Connection connection) throws SQLException {
+        Optional<ApplicationEntity> inserted = JdbcApplicationDao.insertApplication(
+                inputApp,
+                connection);
 
         assertThat(inserted).isNotNull();
         assertThat(inserted.isPresent()).isTrue();
         assertThat(inserted.get()).isEqualTo(inputApp);
+        return null;
     }
 }
