@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-public class H2Configuration {
+public class H2ServletConfig {
 
     @Bean
     ServletRegistrationBean<WebServlet> provideH2ServletRegistrationBean() {
