@@ -1,22 +1,34 @@
 package axal25.oles.jacek.dao.ticket;
 
 import axal25.oles.jacek.entity.TicketEntity;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import javax.transaction.Transactional;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 
-@Repository("springCrudRepository")
-public interface TicketSpringCrudRepository extends CrudRepository<TicketEntity, Integer>, ITicketDao {
+@Repository("crudRepository")
+public interface TicketCrudRepository extends
+        CrudRepository<TicketEntity, Integer>,
+        ITicketDao,
+        TicketPersistenceContextRepository {
 
     @Override
     default List<TicketEntity> getAllTickets() {
         return StreamSupport.stream(findAll().spliterator(), false)
                 .collect(Collectors.toList());
     }
+
+    @Override
+    @Query("SELECT DISTINCT ticket from TicketEntity ticket" +
+            " LEFT JOIN FETCH ticket.release release" +
+            " LEFT JOIN FETCH release.applications apps" +
+            " ORDER BY ticket.id")
+    List<TicketEntity> getAllTicketsEagerly();
 
     @Override
     default TicketEntity addTicket(TicketEntity ticket) {
@@ -30,12 +42,21 @@ public interface TicketSpringCrudRepository extends CrudRepository<TicketEntity,
 
     @Transactional
     @Override
+    default TicketEntity getTicketByIdEagerly(int ticketId) {
+        Optional<TicketEntity> optionalTicket = findById(ticketId);
+        optionalTicket.ifPresent(ticket -> {
+            int unusedToFetchEagerly = ticket.getRelease().getApplications().size();
+        });
+        return optionalTicket.orElse(null);
+    }
+
+
+    @Transactional
+    @Override
     default void updateTicket(TicketEntity updated) {
         TicketEntity existing = getTicketById(updated.getId());
         existing.setDescription(updated.getDescription());
-        existing.setApplication(updated.getApplication());
         existing.setTitle(updated.getTitle());
-
     }
 
     @Transactional
