@@ -31,9 +31,19 @@ public class TicketNonJtaDao implements ITicketDao {
 
     @Override
     public List<TicketEntity> getAllTickets() {
-        String jpqlQuery = "SELECT ticket FROM " +
+        String jpqlQuery = "SELECT DISTINCT ticket FROM " +
                 TicketEntity.class.getSimpleName() +
                 " ticket ORDER BY ticket.id";
+        return entityManager.createQuery(jpqlQuery, TicketEntity.class).getResultList();
+    }
+
+    @Override
+    public List<TicketEntity> getAllTicketsEagerly() {
+        String jpqlQuery = "SELECT DISTINCT ticket from " +
+                TicketEntity.class.getSimpleName() +
+                " ticket LEFT JOIN FETCH ticket.release release" +
+                " LEFT JOIN FETCH release.applications apps" +
+                " ORDER BY ticket.id";
         return entityManager.createQuery(jpqlQuery, TicketEntity.class).getResultList();
     }
 
@@ -60,13 +70,19 @@ public class TicketNonJtaDao implements ITicketDao {
     }
 
     @Override
+    public TicketEntity getTicketByIdEagerly(int ticketId) {
+        TicketEntity ticket = entityManager.find(TicketEntity.class, ticketId);
+        int unusedToFetchEagerly = ticket.getRelease().getApplications().size();
+        return ticket;
+    }
+
+    @Override
     public void updateTicket(TicketEntity updated) {
         EntityTransaction entityTransaction = entityManager.getTransaction();
         entityTransaction.begin();
 
         TicketEntity existing = getTicketById(updated.getId());
         existing.setDescription(updated.getDescription());
-        existing.setApplication(updated.getApplication());
         existing.setTitle(updated.getTitle());
 
         try {
