@@ -1,18 +1,13 @@
 package axal25.oles.jacek.dao.ticket;
 
-import axal25.oles.jacek.dao.DaoTestHelper;
-import axal25.oles.jacek.entity.ApplicationEntity;
 import axal25.oles.jacek.entity.TicketEntity;
 import axal25.oles.jacek.entity.factory.EntityFactory;
 import axal25.oles.jacek.entity.factory.TicketEntityFactory;
-import org.apache.commons.lang3.StringUtils;
 import org.junit.jupiter.api.Test;
 
 import java.lang.reflect.Method;
 import java.util.Arrays;
 import java.util.List;
-import java.util.NoSuchElementException;
-import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Collectors;
 
 import static com.google.common.truth.Truth.assertThat;
@@ -65,78 +60,139 @@ public interface ITicketDaoTester {
 
     ITicketDao getTicketDao();
 
-    DaoTestHelper getDaoTestHelper();
+    Class<?> getTestClass();
 
-    default void updateTicket() {
+    default void updateTicket_updatesDescriptionAndTitle() {
+        TicketEntity inputTicket = TicketEntityFactory.produce(
+                "updateTicket_updatesDescriptionAndTitle",
+                getTestClass(),
+                null,
+                null,
+                null,
+                EntityFactory.IdGenerateMode.NULL);
+        String descriptionPrefix = inputTicket.getDescription();
+        String titlePrefix = inputTicket.getTitle();
 
+        inputTicket.setDescription(descriptionPrefix + " before update");
+        inputTicket.setTitle(titlePrefix + " before update");
+
+        addAndAssertTicket(inputTicket);
+
+        inputTicket.setDescription(descriptionPrefix + " after update");
+        inputTicket.setTitle(titlePrefix + " after update");
+
+        getTicketDao().updateTicket(inputTicket);
+
+        TicketEntity retrievedTicket = getTicketDao().getTicketById(inputTicket.getId());
+
+        assertThat(retrievedTicket.getDescription()).isEqualTo(descriptionPrefix + " after update");
+        assertThat(retrievedTicket.getTitle()).isEqualTo(titlePrefix + " after update");
     }
 
-    default void closeTicket() {
+    default void closeTicket_updatesStatusToResolvedValue() {
+        TicketEntity inputTicket = TicketEntityFactory.produce(
+                "closeTicket_updatesStatusToResolvedValue",
+                getTestClass(),
+                null,
+                null,
+                null,
+                EntityFactory.IdGenerateMode.NULL);
+        inputTicket.setStatus("Not-Resolved");
+        addAndAssertTicket(inputTicket);
 
+        getTicketDao().closeTicketById(inputTicket.getId());
+
+        TicketEntity retrievedTicket = getTicketDao().getTicketById(inputTicket.getId());
+        assertThat(retrievedTicket.getStatus()).isEqualTo("Resolved");
     }
 
     default void deleteTicket_deletesAddedTicket() {
         TicketEntity inputTicket = TicketEntityFactory.produce(
                 "deleteTicket_deletesAddedTicket",
-                TicketCrudRepositoryTest.class,
+                getTestClass(),
                 null,
                 null,
                 null,
                 EntityFactory.IdGenerateMode.NULL);
+        addAndAssertTicket(inputTicket);
 
-        getDaoTestHelper().wrapInTransaction(entityManager -> {
-            addAndAssertTicket(inputTicket);
-            getTicketByIdAndAssertIsEqualTo(inputTicket);
+        getTicketDao().deleteTicketById(inputTicket.getId());
 
-            getTicketDao().deleteTicketById(inputTicket.getId());
-
-            TicketEntity retrievedTicket = getTicketDao().getTicketById(inputTicket.getId());
-            assertThat(retrievedTicket).isNull();
-        });
+        TicketEntity retrievedTicket = getTicketDao().getTicketById(inputTicket.getId());
+        assertThat(retrievedTicket).isNull();
     }
 
-    default void getAllTickets_containsAddedTicket() {
+    default void getAllTickets_containsAddedTicket_skippingLazilyFetchedEntities() {
         TicketEntity inputTicket = TicketEntityFactory.produce(
-                "getAllTickets_containsAddedTicket",
-                TicketCrudRepositoryTest.class,
+                "getAllTickets_containsAddedTicket_skippingLazilyFetchedEntities",
+                getTestClass(),
                 null,
                 null,
                 null,
                 EntityFactory.IdGenerateMode.NULL);
+        addAndAssertTicket(inputTicket);
 
-        AtomicReference<List<TicketEntity>> allTicketsReference = new AtomicReference<>();
+        List<TicketEntity> allTickets = getTicketDao().getAllTickets();
 
-        getDaoTestHelper().wrapInTransaction(entityManager -> {
-            addAndAssertTicket(inputTicket);
-            List<TicketEntity> allTickets = getTicketDao().getAllTickets();
-            allTickets.forEach(ticket -> {
-                boolean unusedToFetchLazyInitializations = ticket.getRelease().getApplications().isEmpty();
-            });
-            allTicketsReference.set(allTickets);
-        });
+        allTickets = allTickets.stream()
+                .map(ticket -> ticket.toBuilder()
+                        .release(ticket.getRelease() == null
+                                ? null
+                                : ticket.getRelease().toBuilder()
+                                .applications(null)
+                                .build())
+                        .build())
+                .collect(Collectors.toList());
+        inputTicket = inputTicket.toBuilder()
+                .release(inputTicket.getRelease() == null
+                        ? null
+                        : inputTicket.getRelease().toBuilder()
+                        .applications(null)
+                        .build())
+                .build();
 
-        assertThat(allTicketsReference.get()).containsAtLeastElementsIn(List.of(inputTicket));
+        assertThat(allTickets).containsAtLeastElementsIn(List.of(inputTicket));
     }
 
-    default void getTicketById_IsEqualToAddedTicket_includingLazyFetchedEntities() {
+    default void getAllTicketsEagerly_containsAddedTicket_includingLazilyFetchedEntities() {
         TicketEntity inputTicket = TicketEntityFactory.produce(
-                "getTicketById_IsEqualToAddedTicket_includingLazyFetchedEntities",
-                TicketCrudRepositoryTest.class,
+                "getAllTicketsEagerly_containsAddedTicket_includingLazilyFetchedEntities",
+                getTestClass(),
+                null,
+                null,
+                null,
+                EntityFactory.IdGenerateMode.NULL);
+        addAndAssertTicket(inputTicket);
+
+        List<TicketEntity> allTickets = getTicketDao().getAllTicketsEagerly();
+
+        assertThat(allTickets.stream()
+                .map(TicketEntity::toClonedComparableFullyFetchedEntity)
+                .collect(Collectors.toList()))
+                .containsAtLeastElementsIn(List.of(inputTicket.toClonedComparableFullyFetchedEntity()));
+    }
+
+    default void getTicketByIdEagerly_IsEqualToAddedTicket_includingLazilyFetchedEntities() {
+        TicketEntity inputTicket = TicketEntityFactory.produce(
+                "getTicketByIdEagerly_IsEqualToAddedTicket_includingLazilyFetchedEntities",
+                getTestClass(),
                 null,
                 null,
                 null,
                 EntityFactory.IdGenerateMode.NULL);
 
         addAndAssertTicket(inputTicket);
-        getDaoTestHelper().wrapInTransaction(entityManager -> {
-            getTicketByIdAndAssertIsEqualTo(inputTicket);
-        });
+
+        TicketEntity retrievedTicket = getTicketDao().getTicketByIdEagerly(inputTicket.getId());
+
+        assertThat(inputTicket.toClonedComparableFullyFetchedEntity())
+                .isEqualTo(retrievedTicket.toClonedComparableFullyFetchedEntity());
     }
 
-    default void getTicketById_isEqualToAddedTicket_skippingLazyFetchedEntities() {
+    default void getTicketById_isEqualToAddedTicket_skippingLazilyFetchedEntities() {
         TicketEntity inputTicket = TicketEntityFactory.produce(
-                "getTicketById_isEqualToAddedTicket_skippingLazyFetchedEntities",
-                TicketCrudRepositoryTest.class,
+                "getTicketById_isEqualToAddedTicket_skippingLazilyFetchedEntities",
+                getTestClass(),
                 null,
                 null,
                 null,
@@ -155,88 +211,74 @@ public interface ITicketDaoTester {
     default void addTicket_isSuccessfullyAdded() {
         TicketEntity inputTicket = TicketEntityFactory.produce(
                 "addTicket_isSuccessfullyAdded",
-                TicketCrudRepositoryTest.class,
+                getTestClass(),
                 null,
                 null,
                 null,
                 EntityFactory.IdGenerateMode.NULL);
-        inputTicket.setId(null);
-        inputTicket.getApplication().setId(null);
-        inputTicket.getRelease().setId(null);
 
         addAndAssertTicket(inputTicket);
     }
 
-    private void getTicketByIdAndAssertIsEqualTo(TicketEntity inputTicket) {
-        TicketEntity retrievedTicket = getTicketDao().getTicketById(inputTicket.getId());
-        assertThat(retrievedTicket).isNotNull();
-        assertThat(retrievedTicket.getRelease()).isNotNull();
+    private void addAndAssertTicket(TicketEntity input) {
+        TicketEntity comparableUnmanaged = input.toClonedComparableFullyFetchedEntity();
+        hasNoIds(input);
 
-        assertThat(inputTicket.toComparableFullyFetchedEntity())
-                .isEqualTo(retrievedTicket.toComparableFullyFetchedEntity());
+        TicketEntity inserted = getTicketDao().addTicket(input);
+
+        hasIds(inserted);
+        hasIds(input);
+        assertThat(inserted).isEqualTo(input);
+        hasNoIds(comparableUnmanaged);
+        assertEqualsIgnoringIds(input, comparableUnmanaged);
     }
 
-    private void addAndAssertTicket(TicketEntity inputTicket) {
-        TicketEntity comparableTicket = inputTicket.toComparableFullyFetchedEntity();
-
-        addAndAssertTicketIds(inputTicket);
-
-        assertManagedAddedAndPojoComparableFullyFetchedTicketAreEqual(inputTicket, comparableTicket);
+    private void hasNoIds(TicketEntity ticket) {
+        assertThat(ticket).isNotNull();
+        assertThat(ticket.getId()).isNull();
+        assertThat(ticket.getApplication()).isNotNull();
+        assertThat(ticket.getApplication().getId()).isNull();
+        assertThat(ticket.getRelease()).isNotNull();
+        assertThat(ticket.getRelease().getId()).isNull();
+        ticket.getRelease().getApplications()
+                .forEach(application -> assertThat(application.getId()).isNull());
+        assertThat(ticket.getRelease().getApplications())
+                .containsAtLeastElementsIn(List.of(ticket.getApplication()));
     }
 
-    private void assertManagedAddedAndPojoComparableFullyFetchedTicketAreEqual(
-            TicketEntity addedManagedTicket,
-            TicketEntity comparableTicket) {
-        // fill Ids on comparable
-        comparableTicket.setId(addedManagedTicket.getId());
-        comparableTicket.getApplication().setId(addedManagedTicket.getApplication().getId());
-        comparableTicket.getRelease().setId(addedManagedTicket.getRelease().getId());
-        comparableTicket.getRelease().getApplications().forEach(comparableApp -> {
-            List<ApplicationEntity> matchingInputApps = addedManagedTicket.getRelease()
-                    .getApplications().stream().filter(inputApp ->
-                            StringUtils.equals(comparableApp.getName(), inputApp.getName())
-                                    && StringUtils.equals(comparableApp.getDescription(), inputApp.getDescription())
-                                    && StringUtils.equals(comparableApp.getOwner(), inputApp.getOwner()))
-                    .collect(Collectors.toList());
-            if (matchingInputApps.size() != 1) {
-                throw new NoSuchElementException("Could not find matching " +
-                        ApplicationEntity.class.getSimpleName() +
-                        "s. Comparable " +
-                        ApplicationEntity.class.getSimpleName() +
-                        ": " +
-                        comparableApp +
-                        ", Matching Added Managed " +
-                        ApplicationEntity.class.getSimpleName() +
-                        " [" +
-                        matchingInputApps.stream().map(Object::toString).collect(Collectors.joining(", ")) +
-                        "].");
-            }
-            comparableApp.setId(matchingInputApps.get(0).getId());
-        });
-
-        assertThat(comparableTicket).isEqualTo(addedManagedTicket.toComparableFullyFetchedEntity());
+    private void hasIds(TicketEntity ticket) {
+        assertThat(ticket).isNotNull();
+        assertThat(ticket.getId()).isNotNull();
+        assertThat(ticket.getApplication()).isNotNull();
+        assertThat(ticket.getApplication().getId()).isNotNull();
+        assertThat(ticket.getRelease()).isNotNull();
+        assertThat(ticket.getRelease().getId()).isNotNull();
+        ticket.getRelease().getApplications()
+                .forEach(application -> assertThat(application.getId()).isNotNull());
+        assertThat(ticket.getRelease().getApplications())
+                .containsAtLeastElementsIn(List.of(ticket.getApplication()));
     }
 
-    private void addAndAssertTicketIds(TicketEntity inputTicket) {
-        assertThat(inputTicket).isNotNull();
-        assertThat(inputTicket.getApplication()).isNotNull();
-        assertThat(inputTicket.getRelease()).isNotNull();
-        assertThat(inputTicket.getId()).isNull();
-        assertThat(inputTicket.getApplication().getId()).isNull();
-        assertThat(inputTicket.getRelease().getId()).isNull();
-
-        TicketEntity inserted = getTicketDao().addTicket(inputTicket);
-
-        assertThat(inserted).isNotNull();
-        assertThat(inserted).isEqualTo(inputTicket);
-
-        assertThat(inserted.getId()).isNotNull();
-        assertThat(inserted.getApplication().getId()).isNotNull();
-        assertThat(inserted.getRelease().getId()).isNotNull();
-
-        inserted.getRelease().getApplications().forEach(application ->
-                assertThat(application.getId()).isNotNull());
-        assertThat(inserted.getRelease().getApplications())
-                .containsAtLeastElementsIn(List.of(inserted.getApplication()));
+    private void assertEqualsIgnoringIds(
+            TicketEntity addedManaged,
+            TicketEntity comparableUnmanaged) {
+        // .toClonedComparableFullyFetchedEntity()
+        // needs to be re-sorted because it got newly issued ids
+        assertThat(comparableUnmanaged)
+                .isEqualTo(addedManaged.toBuilder()
+                        .id(null)
+                        .application(addedManaged.getApplication().toBuilder()
+                                .id(null)
+                                .build())
+                        .release(addedManaged.getRelease().toBuilder()
+                                .id(null)
+                                .applications(addedManaged.getRelease().getApplications().stream()
+                                        .map(application -> application.toBuilder()
+                                                .id(null)
+                                                .build())
+                                        .collect(Collectors.toList()))
+                                .build())
+                        .build()
+                        .toClonedComparableFullyFetchedEntity());
     }
 }
