@@ -1,6 +1,7 @@
 package axal25.oles.jacek.port;
 
-import axal25.oles.jacek.config.AppInfoListener;
+import axal25.oles.jacek.context.info.BugTrackerAppStatusProducerOnApplicationReady;
+import axal25.oles.jacek.model.BugTrackerAppStatus;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -13,12 +14,14 @@ import static com.google.common.truth.Truth.assertThat;
 public class DefinedPortTest {
 
     @Autowired
-    private AppInfoListener appInfoListener;
+    private BugTrackerAppStatusProducerOnApplicationReady bugTrackerAppStatusProducerOnApplicationReady;
 
     @Test
     void isPortNull() {
-        assertThat(appInfoListener.getPort()).isEqualTo(8090);
-        assertThat(appInfoListener.getLocal().getPort()).isEqualTo(8090);
-        assertThat(appInfoListener.getRemote().getPort()).isNull();
+        BugTrackerAppStatus bugTrackerAppStatus =
+                bugTrackerAppStatusProducerOnApplicationReady.getBugTrackerAppStatusProducer().getBugTrackerAppStatus();
+        assertThat(bugTrackerAppStatus.getServerDotPort()).isEqualTo(8090);
+        assertThat(bugTrackerAppStatus.getLocal().getPort()).isEqualTo(8090);
+        assertThat(bugTrackerAppStatus.getRemote().getPort()).isNull();
     }
 }
